Added showing cell's units, selecting, sending units

+++ core/src/com/github/piotrkruk/phage_wars/model/Race.java
@@ -8,17 +8,17 @@ package com.github.piotrkruk.phage_wars.model;
 
 public class Race {
 	
-	final int growthRate;
+	final double growthRate;
 	final double givesAway;
 	
 	// default values for race's parameters
-	final static int DEFAULT_GROWTH = 5;
+	final static double DEFAULT_GROWTH = 5;
 	final static double DEFAULT_GIVES = 0.3;
 	
 	// each race has it's owner
 	final Player owner;
 	
-	public Race(int growth, double gives, Player owner) {
+	public Race(double growth, double gives, Player owner) {
 		growthRate = growth;
 		givesAway = gives;
 		
@@ -29,7 +29,7 @@ public class Race {
 		this(DEFAULT_GROWTH, DEFAULT_GIVES, owner);
 	}
 	
-	int givesAway(int units) {
-		return (int) givesAway * units;
+	double givesAway(double units) {
+		return givesAway * units;
 	}
 }
